# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Signal Audit API",
    "description": "Measures how many independent signals a set of metrics actually carries, and which columns can be removed without losing information.\n\n**Headline figures use first-differenced series.** Metrics that all grow over time correlate from shared trend alone; the raw view is returned alongside so the gap is visible.\n\n**Executable exports are gated.** The engine proves statistical redundancy; it cannot see that a metric is the sole condition on a paging monitor. A completed blast-radius worksheet is required before Terraform or exclusion plans are returned.\n\n**No cost figure is synthesised.** Column redundancy and billing cardinality differ by orders of magnitude, so savings estimates require customer-supplied unit cost and, ideally, a cardinality map.",
    "version": "1.1"
  },
  "paths": {
...
    "/v1/audits/preflight": {
      "post": {
        "tags": [
          "audits"
        ],
        "summary": "Shape, quota verdict and execution mode without analysing",
        "description": "Shape, quota verdict and execution mode WITHOUT running the\nanalysis. Lets the client show limits and set expectations before\ncommitting to a long job.",
        "operationId": "preflight_v1_audits_preflight_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preflight_v1_audits_preflight_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResponse"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/audits": {
      "post": {
        "tags": [
          "audits"
        ],
        "summary": "Run an audit (inline when small, queued when large)",
        "description": "Run an audit.\n\nmode=auto (default) runs inline when small and queues when large,\nper the tier's threshold. The response always states which happened,\nso a client never has to infer it from the shape of the body.\n\nBasis declaration. `basis` names which lens the headline reports:\nraw, differenced, or ratio:COL. `scale_by` is a comma-separated list\nof denominators, each adding one ratio basis; `scale_exempt` names\ncolumns that are ALREADY scale-free and must not be divided \u2014\ndividing a rate by a size injects the very confound the ratio basis\nremoves. All optional: omitting them reproduces the previous\nresponse exactly, and the contract reports `basis.declared: false`\nso a client can enforce what the server does not.\n\n`require_basis=true` makes the server refuse an undeclared request.\nIt is opt-in rather than default because defaulting it on would\nbreak every existing caller.",
        "operationId": "create_audit_v1_audits_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_audit_v1_audits_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncAuditResponse"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncAuditResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payment Required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      },
      "get": {
        "tags": [
          "audits"
        ],
        "summary": "List this account's audits",
        "operationId": "list_audits_v1_audits_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audits/{job_id}": {
      "get": {
        "tags": [
          "audits"
        ],
        "summary": "Job status, and the contract once complete",
        "operationId": "get_audit_v1_audits__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/usage": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Consumption against the tier allowance",
        "description": "Consumption for the current account. Sums the metering unit\nacross completed audits; a real deployment reads this from the\nbilling store rather than from job memory.",
        "operationId": "usage_v1_usage_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audits/{job_id}/exports": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "What can be downloaded now, and what is gated",
        "description": "What can be downloaded now, and what is still gated behind the\nblast-radius attestation.",
        "operationId": "list_exports_v1_audits__job_id__exports_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportAvailability"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audits/{job_id}/exports/review-worksheet": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Blast-radius worksheet (CSV) \u2014 request this first",
        "description": "The artefact that should be requested first: statistical evidence\npre-filled, operational columns blank for the operator.",
        "operationId": "export_worksheet_v1_audits__job_id__exports_review_worksheet_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audits/{job_id}/exports/column-manifest": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Keep/drop decision per column (CSV), ungated",
        "description": "Keep/drop per column. A document, not a change \u2014 ungated.",
        "operationId": "export_manifest_v1_audits__job_id__exports_column_manifest_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_preflight_v1_audits_preflight_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "ignore": {
            "type": "string",
            "title": "Ignore",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preflight_v1_audits_preflight_post"
      },
...
      "PreflightResponse": {
        "properties": {
          "metrics": {
            "type": "integer",
            "title": "Metrics"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "cells": {
            "type": "integer",
            "title": "Cells"
          },
          "rows_per_metric": {
            "type": "number",
            "title": "Rows Per Metric"
          },
          "within_quota": {
            "type": "boolean",
            "title": "Within Quota"
          },
          "quota_error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Error"
          },
          "recommended_mode": {
            "type": "string",
            "enum": [
              "sync",
              "async"
            ],
            "title": "Recommended Mode"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "metrics",
          "rows",
          "cells",
          "rows_per_metric",
          "within_quota",
          "recommended_mode",
          "tier"
        ],
        "title": "PreflightResponse"
      },
...
    }
  },
...
}